improvement(bluesky): removed blueskysender class
